@@ -1,14 +1,12 @@
 #include <Duck.h>
 
-Duck::Duck() = default;
+void Duck::swim() const {
+  std::cout << "All ducks float, even decoys!" << std::endl;
+}
 
-Duck::~Duck() = default;
+void Duck::performFly() const { flyBehaviour->fly(); }
 
-void Duck::swim() { std::cout << "All ducks float, even decoys!" << std::endl; }
-
-void Duck::performFly() { flyBehaviour->fly(); }
-
-void Duck::performQuack() { quackBehaviour->quack(); }
+void Duck::performQuack() const { quackBehaviour->quack(); }
 
 void Duck::setFlyBehaviour(std::unique_ptr<FlyBehaviour> fb) {
   flyBehaviour = std::move(fb);
