@@ -1,7 +1,3 @@
 #include <FlyNoWay.h>
 
-FlyNoWay::FlyNoWay() = default;
-
-FlyNoWay::~FlyNoWay() = default;
-
-void FlyNoWay::fly() { std::cout << "I can't fly" << std::endl; }
+void FlyNoWay::fly() const { std::cout << "I can't fly" << std::endl; }
