@@ -1,7 +1,3 @@
 #include <FlyWithWings.h>
 
-FlyWithWings::FlyWithWings() = default;
-
-FlyWithWings::~FlyWithWings() = default;
-
-void FlyWithWings::fly() { std::cout << "I'm flying!!" << std::endl; }
+void FlyWithWings::fly() const { std::cout << "I'm flying!!" << std::endl; }
