@@ -1,7 +1,3 @@
 #include <Quack.h>
 
-Quack::Quack() = default;
-
-Quack::~Quack() = default;
-
-void Quack::quack() { std::cout << "Quack" << std::endl; }
+void Quack::quack() const { std::cout << "Quack" << std::endl; }
